refactor:: 컴포넌트 쪼개기, 검색 최적화 - 장황한 컴포넌트들을 기능에 따라 파일에 분류해두었습니다 - useQuery에서 제공하는 refetch를 이용해 검색을 최적화하였습니다
