+++ src/components/timetable/Grid/LectureSticker.tsx
