--- src/util/generateRandomString.ts
@@ -1,10 +1,0 @@
-export const generateRandomString = (num: number) => {
-  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'
-  let result = ''
-  const charactersLength = characters.length
-  for (let i = 0; i < num; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength))
-  }
-
-  return result
-}
