+++ src/util/useCourseSearch.ts
@@ -9,7 +9,6 @@ import {
   getByProfInMajor,
   getInAcademicFoundation,
 } from '@/api/hooks/course'
-import { initCourseSearchData } from '@/util/timetableUtil'
 
 export interface useCourseSearchProps {
   queryKeyword: string
@@ -21,7 +20,7 @@ export interface useCourseSearchProps {
 export const useCourseSearch = ({ queryKeyword, category, classification, filter }: useCourseSearchProps) => {
   const authHeader = useAuthHeader()
 
-  const { data } = useQuery({
+  const { data, refetch: search } = useQuery({
     queryKey: ['courseSearchResult', queryKeyword],
     queryFn: () => {
       if (category === 'Academic Foundations') {
@@ -49,10 +48,10 @@ export const useCourseSearch = ({ queryKeyword, category, classification, filter
         return getByProfInMajor({ authHeader, professorName: queryKeyword, major: classification! })
       }
     },
-    enabled: !!queryKeyword,
+    enabled: false,
     retry: false,
-    initialData: initCourseSearchData,
+    initialData: { hasNextPage: false, nextCursorId: 0, data: [] },
   })
 
-  return data
+  return { data, search }
 }
